+++ orttraining/orttraining/core/graph/optimizer_builder.h
@@ -95,11 +95,11 @@ class OptimizerBuilder {
    * @param config The input config for optimizer builder
    * @param[out] graph_defs The GraphDefs corresponding to the graph (possibly
    *             a subgraph) that the component is to be added to.
-   * @param[out] weight_to_opt_mapping Mapping between weight to corresponding optimizer states. Includes 
-   *             all initializers that should be
+   * @param[out] weight_to_opt_mapping Mapping between weight to corresponding optimizer states. 
+   *             Includes all initializers that should be
    *             placed in the parent graph, if there is one.
    *             Other initializers are treated as local to the current
-   *             (sub)graph.
+   *             (sub)graph. These are new / external states added for optimizer.
    * @param[out] output_weight_argdefs The output weight ArgDef. All optimizers
                  should have this output.
    * @param[out] output_gradient_argdefs The output gradient ArgDef. All optimizers
